feat(communications): Cabin Intercommunication Data System
This commit introduces the first version of a Cabin Intercommunication
Data System (CIDS) for the A32NX, focussed around properly simualing
passenger signs in the cabin and their associated chimes. It also
implements the cabin ready functionality, although for the time being
this will still have to be set by the pilot.

A future cabin simulation can hook into the signs to enable realistic
reactions and cabin ready timings.